+++ Lesson5/src/main/java/comm/tms/additional/Task1/objectTask1/TypeOfHdd.java
@@ -1,0 +1,6 @@
+package comm.tms.additional.Task1.objectTask1;
+
+public enum TypeOfHdd {
+    EXTERNAL,
+    INTERNAL;
+}
